+++ dog/dog/out.h
@@ -1,11 +1,12 @@
 #pragma once
 #include <string>
 #include <iostream>
+#include <iomanip>
 
 class out
 {
 public:
 	out(std::string toOutput);
-	out(int toOutput);
+	out(double toOutput);
 	~out();
 };
